ref(frontend): Refactor useApi to the /api folder.
