+++ frontend/src/components/PageSettings/PageSettings.tsx
@@ -7,7 +7,7 @@ import Icon from "../Icon/Icon"
 import Spacer from "../Spacer/Spacer"
 import { ReactComponent as ShareIcon } from "../../icons/share.svg"
 import Button from "../Button/Button"
-import useApi from "../../hooks/useApi"
+import useApi from "../../api/useApi"
 import { getSchoolId } from "../../hooks/schoolIdState"
 import UserCard from "../UserCard/UserCard"
 import CardLink from "../CardLink/CardLink"
